enable connecting to out-of-process kernels over stdio
