+++ dotnet-interactive.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31603.294
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31709.452
 MinimumVisualStudioVersion = 15.0.26124.0
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "dotnet-interactive", "src\dotnet-interactive\dotnet-interactive.csproj", "{1A4A744F-FEF3-4361-A2E5-C5D496F7BFDC}"
 EndProject
@@ -44,10 +44,6 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DotNet.Interactive.PowerShell", "src\Microsoft.DotNet.Interactive.PowerShell\Microsoft.DotNet.Interactive.PowerShell.csproj", "{CE412E28-8660-43CA-8E0B-293EAC624623}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DotNet.Interactive.PowerShell.Tests", "src\Microsoft.DotNet.Interactive.PowerShell.Tests\Microsoft.DotNet.Interactive.PowerShell.Tests.csproj", "{30DE6920-EA82-4E4F-BFCA-1FF179FAA2DA}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "dotnet-interactive.IntegrationTests", "src\dotnet-interactive.IntegrationTests\dotnet-interactive.IntegrationTests.csproj", "{8017943B-B591-49C8-B957-40141B427AD9}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "integration-tests", "integration-tests", "{979A5022-09CA-4258-B76D-CFDC31DA71E0}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "interface-generator", "src\interface-generator\interface-generator.csproj", "{9A091E80-D5EB-4DEF-8E38-B24B128393DF}"
 EndProject
@@ -326,18 +322,6 @@ Global
 		{30DE6920-EA82-4E4F-BFCA-1FF179FAA2DA}.Release|x64.Build.0 = Release|Any CPU
 		{30DE6920-EA82-4E4F-BFCA-1FF179FAA2DA}.Release|x86.ActiveCfg = Release|Any CPU
 		{30DE6920-EA82-4E4F-BFCA-1FF179FAA2DA}.Release|x86.Build.0 = Release|Any CPU
-		{8017943B-B591-49C8-B957-40141B427AD9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8017943B-B591-49C8-B957-40141B427AD9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8017943B-B591-49C8-B957-40141B427AD9}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{8017943B-B591-49C8-B957-40141B427AD9}.Debug|x64.Build.0 = Debug|Any CPU
-		{8017943B-B591-49C8-B957-40141B427AD9}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{8017943B-B591-49C8-B957-40141B427AD9}.Debug|x86.Build.0 = Debug|Any CPU
-		{8017943B-B591-49C8-B957-40141B427AD9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8017943B-B591-49C8-B957-40141B427AD9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8017943B-B591-49C8-B957-40141B427AD9}.Release|x64.ActiveCfg = Release|Any CPU
-		{8017943B-B591-49C8-B957-40141B427AD9}.Release|x64.Build.0 = Release|Any CPU
-		{8017943B-B591-49C8-B957-40141B427AD9}.Release|x86.ActiveCfg = Release|Any CPU
-		{8017943B-B591-49C8-B957-40141B427AD9}.Release|x86.Build.0 = Release|Any CPU
 		{9A091E80-D5EB-4DEF-8E38-B24B128393DF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{9A091E80-D5EB-4DEF-8E38-B24B128393DF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9A091E80-D5EB-4DEF-8E38-B24B128393DF}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -482,7 +466,6 @@ Global
 		{499F936C-AE39-477B-821A-FF48F9E1A0D0} = {11BA3480-4584-435C-BA9A-8C554DB60E9F}
 		{CE412E28-8660-43CA-8E0B-293EAC624623} = {B95A8485-8C53-4F56-B0CE-19C0726B5805}
 		{30DE6920-EA82-4E4F-BFCA-1FF179FAA2DA} = {11BA3480-4584-435C-BA9A-8C554DB60E9F}
-		{8017943B-B591-49C8-B957-40141B427AD9} = {979A5022-09CA-4258-B76D-CFDC31DA71E0}
 		{9A091E80-D5EB-4DEF-8E38-B24B128393DF} = {B95A8485-8C53-4F56-B0CE-19C0726B5805}
 		{DE942DC3-B41B-45C5-9E01-3C4F042D3916} = {B95A8485-8C53-4F56-B0CE-19C0726B5805}
 		{5167FC61-2FBF-4E0D-877E-C2DD354FB955} = {B95A8485-8C53-4F56-B0CE-19C0726B5805}
