+++ src/dotnet-interactive-vscode/common/stdioKernelTransport.ts
@@ -114,7 +114,7 @@ export class StdioKernelTransport implements KernelTransport {
         }
         catch (error) {
             this.diagnosticChannel.appendLine(`Failure setting up tunnel confinguration for Kernel process ${this.childProcess?.pid}`);
-            this.diagnosticChannel.appendLine(`Error : ${error.message}`);
+            this.diagnosticChannel.appendLine(`Error : ${(<any>error)?.message}`);
             return null;
         }
     }
