--- src/dotnet-interactive.IntegrationTests/Directory.Build.props
@@ -1,9 +1,0 @@
-<Project>
-
-  <PropertyGroup>
-    <UseBetaVersion>true</UseBetaVersion>
-  </PropertyGroup>
-
-  <Import Project="$([MSBuild]::GetPathOfFileAbove('Directory.Build.props', '$(MSBuildThisFileDirectory)../'))" />
-
-</Project>
