--- src/dotnet-interactive.IntegrationTests/README.md
@@ -1,1 +1,0 @@
-The unit tests in this project are run as a special leg of CI that assumes the `dotnet interactive` global tool has been installed.
