--- src/dotnet-interactive.IntegrationTests/Utility/ProcessHelper.cs
@@ -1,65 +1,0 @@
-// Copyright (c) .NET Foundation and contributors. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-
-using System;
-using System.Diagnostics;
-using System.IO;
-using System.Text;
-
-namespace Microsoft.DotNet.Interactive.App.IntegrationTests.Utility
-{
-    public static class ProcessHelper
-    {
-        public static Process Start(
-            string command,
-            string args,
-            DirectoryInfo workingDir,
-            Action<string> output = null,
-            Action<string> error = null)
-        {
-            var process = new Process
-            {
-                StartInfo =
-                    {
-                        Arguments = args,
-                        FileName = command,
-                        RedirectStandardError = true,
-                        RedirectStandardOutput = true,
-                        RedirectStandardInput = true,
-                        WorkingDirectory = workingDir?.FullName ?? string.Empty,
-                        StandardOutputEncoding = Encoding.UTF8
-                    }
-            };
-
-            if (output is not null)
-            {
-                process.OutputDataReceived += (sender, eventArgs) =>
-                {
-                    if (eventArgs.Data is not null)
-                    {
-                        output(eventArgs.Data);
-                    }
-                };
-            }
-
-            if (error is not null)
-            {
-                process.ErrorDataReceived += (sender, eventArgs) =>
-                {
-                    if (eventArgs.Data is not null)
-                    {
-                        error(eventArgs.Data);
-                    }
-                };
-            }
-
-            process.Start();
-
-            process.BeginOutputReadLine();
-            process.BeginErrorReadLine();
-
-            return process;
-        }
-    }
-}
