+++ src/dotnet-interactive.Tests/CommandLine/CommandLineParserTests.cs
@@ -2,21 +2,27 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
+using System.Collections.Generic;
 using System.CommandLine.Binding;
 using System.CommandLine.IO;
 using System.CommandLine.Parsing;
 using System.IO;
 using System.Linq;
+using System.Threading;
 using System.Threading.Tasks;
 
 using FluentAssertions;
 using FluentAssertions.Execution;
 
 using Microsoft.DotNet.Interactive.App.CommandLine;
+using Microsoft.DotNet.Interactive.App.Tests.Extensions;
+using Microsoft.DotNet.Interactive.Commands;
+using Microsoft.DotNet.Interactive.Connection;
 using Microsoft.DotNet.Interactive.Http;
 using Microsoft.DotNet.Interactive.Server;
 using Microsoft.DotNet.Interactive.Telemetry;
 using Microsoft.DotNet.Interactive.Tests.Utility;
+using Microsoft.DotNet.Interactive.Utility;
 using Microsoft.Extensions.DependencyInjection;
 
 using Xunit;
@@ -93,6 +99,38 @@ namespace Microsoft.DotNet.Interactive.App.Tests.CommandLine
                 .FullName
                 .Should()
                 .Be(logPath.FullName);
+        }
+
+        [Fact]
+        public async Task kernel_server_honors_log_path()
+        {
+            using var logPath = DisposableDirectory.Create();
+
+            _output.WriteLine($"Created log file: {logPath.Directory.FullName}");
+
+            var waitTime = TimeSpan.FromSeconds(10);
+
+            using (var kernel = new CompositeKernel().UseKernelClientConnection(new ConnectStdIO()))
+            {
+                await kernel.SendAsync(new SubmitCode($"#!connect stdio --kernel-name proxy --command \"{Dotnet.Path}\" \"{typeof(Program).Assembly.Location}\" stdio --log-path \"{logPath.Directory.FullName}\" --verbose --wait-for-kernel-ready-event true"));
+
+                await kernel.SendAsync(new SubmitCode("1+1", "proxy"));
+            }
+
+            // wait for log file to be created
+            var logFile = await logPath.Directory.WaitForFile(
+                              timeout: waitTime,
+                              predicate: _file => true); // any matching file is the one we want
+            logFile.Should().NotBeNull($"a log file should have been created at {logFile.FullName}");
+
+            // check log file for expected contents
+            (await logFile.WaitForFileCondition(
+                 timeout: waitTime,
+                 predicate: file => file.Length > 0))
+                .Should()
+                .BeTrue($"expected non-empty log file within {waitTime.TotalSeconds}s");
+            var logFileContents = File.ReadAllText(logFile.FullName);
+            logFileContents.Should().Contain("CodeSubmissionReceived: 1+1");
         }
 
         [Fact]
